requirements/base.txt
Django>=1.8
requests>=2.7
pyjwt>=1.4.0

